# services/adapter-fabric/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TANTU Adapter Fabric",
    "description": "Multi-protocol shopfloor ingestion \u2014 OPC-UA, Modbus, MQTT, MTConnect, EtherNet/IP, Camera \u2014 normalized to one schema, derived DefectEvents only (no image field).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ingest": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Direct ingest of a normalized reading (for gateways)",
        "operationId": "ingest_reading_ingest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestReadingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Synthetic derived DefectEvents (demo)",
        "operationId": "list_events_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestReadingIn": {
        "properties": {
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "unknown"
          },
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id",
            "default": ""
          },
          "source_tag": {
            "type": "string",
            "title": "Source Tag",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "station_id",
          "metric",
          "value"
        ],
        "title": "IngestReadingIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}